+++ KrieptoBot.Tests/Application/Recommendators/RecommendatorScoreTests.cs
@@ -8,8 +8,8 @@ namespace KrieptoBot.Tests.Application.Recommendators
         [Test]
         public void RecommendationScore_ShouldDoMath_Correct()
         {
-            const float score1Value = 20;
-            const float score2Value = 60;
+            const decimal score1Value = 20;
+            const decimal score2Value = 60;
             var score1 = new RecommendatorScore { Score = score1Value };
             var score2 = new RecommendatorScore { Score = score2Value };
 
@@ -17,7 +17,7 @@ namespace KrieptoBot.Tests.Application.Recommendators
             Assert.That((score1 / score2).Score, Is.EqualTo( score1Value / score2Value ));
             Assert.That((score1 + score2).Score, Is.EqualTo( score1Value + score2Value ));
             Assert.That((score1 - score2).Score, Is.EqualTo( score1Value - score2Value ));
-                        
+
             Assert.That((score1 * 2).Score, Is.EqualTo(score1Value * 2 ));
             Assert.That((score1 / 2).Score, Is.EqualTo(score1Value / 2 ));
             Assert.That((score1 + 2).Score, Is.EqualTo(score1Value + 2 ));
